backend/requirements.txt
# Core
fastapi==0.104.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
elasticsearch==8.12.0
elasticsearch[async]==8.12.0

# SQL Database (metadados do sistema)
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Encryption (para senhas de ES servers)
cryptography==41.0.7

# Cache & Pub/Sub (opcional)
redis==5.0.1

# WebSocket
python-socketio==5.10.0
python-engineio==4.8.0

# AI & LLM
langchain==0.1.0
langchain-community==0.0.10
langchain-core>=0.1.8,<0.2
anthropic==0.21.3
openai==1.54.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dateutil==2.8.2
httpx==0.25.2

# SSO & OAuth2
authlib==1.3.0
msal==1.28.0
apscheduler==3.10.4

# Export (PDF & Images)
reportlab==4.0.7
playwright==1.40.0
pillow==10.1.0
plotly==5.18.0
kaleido==0.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
